Python automation project
Skillgun is a website maintained by Palle Technologies for the operations of the training institute management which includes several modules like placements, students, employees, operations. The purpose of this project is to automate the test cases of student module. As pat of this project I have automated the student login, edit profile, edit contact page test cases.